+++ modules/v1_8_R2/src/main/java/de/Keyle/MyPet/compat/v1_8_R2/entity/EntityMyPet.java
@@ -433,7 +433,7 @@ public abstract class EntityMyPet extends EntityCreature implements IAnimal, MyP
                             return true;
                         }
                     }
-                    int addHunger = Configuration.HungerSystem.HUNGER_SYSTEM_POINTS_PER_FEED;
+                    double addHunger = Configuration.HungerSystem.HUNGER_SYSTEM_POINTS_PER_FEED;
                     if (getHealth() < getMaxHealth()) {
                         if (!entityhuman.abilities.canInstantlyBuild) {
                             --itemStack.count;
